requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
Pillow==9.5.0
requests==2.31.0
tkhtmlview==0.2.0
urllib3==2.1.0
